Separate the single solution file into separate tasks and their solutions and add an initial cheatsheet of TA-relevant methods and commands
