+++ exercises/tasks/task_01.py
@@ -1,0 +1,21 @@
+# Searching for a connection displays results
+def test_searching_for_connection_displays_results(page):
+    # 1. Open the kiwi.com website (wait for page to load)
+    pass
+    # 2. Clear the `from` location
+
+    # 3. Type in `Vienna` to the `from` field
+
+    # 4. Select the 1st result from the dropdown
+
+    # 5. Type in `Brno` to the `to` field
+
+    # 6. Select the 1st result from the dropdown
+
+    # 7. Uncheck the `Booking` checkbox
+
+    # 8. Hit the `Search` button
+
+    # 9. Available connections should be displayed
+
+    # (10. variation: among the results, this first one is cheaper than 10 000 CZK)
