+++ exercises/tasks/task_02.py
@@ -1,0 +1,30 @@
+# Searching with additional transportation options shows results cheaper than 1 000 CZK
+def test_searching_with_additional_transportation_options_shows_results_cheaper_than_1k_czk(page):
+    # 1. Steps 1-9. from the previous scenario
+    # 1.1. Open the kiwi.com website (wait for page to load)
+    pass
+
+    # 1.2. Clear the `from` location
+
+    # 1.3. Type in `Vienna` to the `from` field
+
+    # 1.4. Select the 1st result from the dropdown
+
+    # 1.5. Type in `Brno` to the `to` field
+
+    # 1.6. Select the 1st result from the dropdown
+
+    # 1.7. Uncheck the `Booking` checkbox
+
+    # 1.8. Hit the `Search` button
+
+    # 1.9. Available connections should be displayed
+
+    # 2. Check the `Bus` checkbox in the `Transport` left-hand section of the results
+
+    # 3. Verify the first result is cheaper than 1 000 CZK
+
+    # (4. variation: on step 2. check the `Train` checkbox in the `Transport` left-hand section as well; on step 3. verify the results are cheaper than 350 CZK)
+    # 4.1 Check the `Train` checkbox in the `Transport` left-hand section as well
+
+    # 4.2 Verify the results are cheaper than 350 CZK
