+++ exercises/tasks/task_08.py
@@ -1,0 +1,25 @@
+# Sidebar actions (expanding options, verifying visibility of items)
+def test_sidebar_actions_work_as_expected(page):
+    # 1. On the Kiwi.com website hit the right sidebar hamburger button
+    # 1.1. Open the kiwi.com website (wait for page to load)
+    pass
+
+    # 1.2. Hit the right sidebar hamburger button
+
+    # 2. Verify a sidebar with the Manage your trips, set up price alerts, use Kiwi.com Credit, and get personalized support. text appears
+
+    # 3. Hit the Discover button
+
+    # 4. Verify the Discover button expands into a dropdown/slide of items
+
+    # 5. Verify the Subscribe to newsletter button is displayed
+
+    # 6. Verify the Stories button is displayed
+
+    # 7. Hit the Subscribe to newsletter button
+
+    # 8. Verify the sidebar disappears
+
+    # 9. Verify a modal with the Subscribe to the Kiwi.com newsletter heading is displayed
+
+    # 10. Verify the modal can be closed by hitting the cross button in its top right corner
