+++ exercises/tasks/task_10.py
@@ -1,0 +1,15 @@
+# Travel mode interactions are respected by the UI
+def test_travel_mode_interactions_are_respected_by_the_ui(page):
+    # 1. On the Kiwi.com website hit the travel mode button (which has the Return value selected by default)
+    # 1.1. Open the kiwi.com website (wait for page to load)
+    pass
+
+    # 1.2. Hit the travel mode button
+
+    # 2. Verify a popup with the following options is displayed: Return, One-way, Multi-city, and Nomad
+
+    # 3. Select the One-way options
+
+    # 4. Verify the One-way option is selected
+
+    # 5. Verify the Return date field is no longer displayed
