+++ exercises/tasks/task_11.py
@@ -1,0 +1,49 @@
+# Total count of passengers is reflected by counter next to the passengers icon
+def test_total_count_of_passengers_is_reflected_by_counter_next_to_the_passengers_icon(page):
+    # 1. On the Kiwi.com website hit the passengers and bags button
+    # 1.1. Open the kiwi.com website (wait for page to load)
+    pass
+
+    # 1.2. Hit the passengers and bags button
+
+    # 2. Verify a popup with the following options is displayed: Adults, Children, Infants in the Passengers section
+
+    # 3. Verify a popup with the following options is displayed: Cabin baggage and Checked baggage in the Bags section
+
+    # 4. Store the values of the count of adults, children and infants into variables
+
+    # 5. Hit twice the + button next to the Adults option
+
+    # 6. Verify the count of adults has increased by 2 (compare with value from step 4.)
+
+    # 7. Hit thrice the + button next to the Children option
+
+    # 8. Verify the count of adults has increased by 3 (compare with value from step 4.)
+
+    # 9. Hit once the + button next to the Infants option
+
+    # 10. Verify the count of adults has increased by 1 (compare with value from step 4.)
+
+    # 11. Hit the Done button
+
+    # 12. Verify the count of passengers next to the passengers icon corresponds with the sum of values from steps 6.,
+    # 8., and 10.
+
+    # (13. variation: increase the count of baggage and verify the counters correspond with the count of added pieces
+    # of baggage)
+    # 13.1. Hit the passengers and bags button
+
+    # 13.2. Store the values of the count of cabin and checked baggage pieces into variables
+
+    # 13.3. Hit thrice the + button next to the Cabin baggage option
+
+    # 13.4. Verify the count of adults has increased by 3 (compare with value from step 13.2.)
+
+    # 13.5. Hit twice the + button next to the Checked baggage option
+
+    # 13.6. Verify the count of adults has increased by 2 (compare with value from step 13.2.)
+
+    # 13.7. Hit the Done button
+
+    # 13.8. Verify the count of passengers next to the passengers icon corresponds with the sum of values from steps
+    # 13.4. and 13.6.
